Route-emitter communicates directly with the bbs
* no longer use receptor to get desired/actual lrps

[#102265600]

+++ routing_table/registry_message.go
@@ -2,7 +2,7 @@ package routing_table
 
 type RegistryMessage struct {
 	Host string   `json:"host"`
-	Port uint16   `json:"port"`
+	Port uint32   `json:"port"`
 	URIs []string `json:"uris"`
 	App  string   `json:"app,omitempty"`
